refactor(backend): reorganize backend structure for improved modularity and maintainability
