+++ src/backend/efficiencies/models.py
@@ -1,0 +1,8 @@
+from typing import List
+from pydantic import BaseModel
+
+
+class EfficiencyRequest(BaseModel):
+    experimentId: str
+    selectedEfficiencies: List[str]
+    timeInterval: int
