+++ src/backend/table/models.py
@@ -1,0 +1,31 @@
+from typing import Dict, List
+from pydantic import BaseModel
+
+
+class DataRequest(BaseModel):
+    experimentId: str
+
+
+class UpdateDataPayload(BaseModel):
+    updatedData: Dict[str, Dict[str, object]]
+
+
+class AddColumnRequest(BaseModel):
+    columnName: str
+    defaultValue: str | int | None = None
+
+
+class AddRowRequest(BaseModel):
+    rowData: dict
+
+
+class RemoveColumnRequest(BaseModel):
+    columnName: str
+
+
+class RemoveRowRequest(BaseModel):
+    experimentIds: List[str]
+
+
+class SetColumnTypes(BaseModel):
+    newColumnTypes: Dict[str, str]
